+++ lib/hcloud/entities/storage_box_stats.rb
@@ -1,0 +1,9 @@
+# frozen_string_literal: true
+
+module HCloud
+  class StorageBoxStats < Entity
+    attribute :size, :integer
+    attribute :size_data, :integer
+    attribute :size_snapshots, :integer
+  end
+end